v0.8.1 Added many S3 methods, fixed printing, added new instructions

+++ inst/include/glmmr/calculator.hpp
@@ -315,7 +315,7 @@ inline dblvec glmmr::calculator::calculate(const int i,
     case Do::PushParameter:
     {
 #if defined(ENABLE_DEBUG) && defined(R_BUILD)
-      if((unsigned)idx_iter >= indexes.size())Rcpp::stop("Index out of range: case 2 idx iter: "+std::to_string(idx_iter)+" versus "+std::to_string(indexes.size()));
+      if(idx_iter >= indexes.size())Rcpp::stop("Index out of range: case 2 idx iter: "+std::to_string(idx_iter)+" versus "+std::to_string(indexes.size()));
       if(indexes[idx_iter] >= parameter_count)Rcpp::stop("Index out of range: case 2 indexes: "+std::to_string(indexes[idx_iter])+" versus "+std::to_string(parameter_count));
       if(indexes[idx_iter] >= parameters.size())Rcpp::stop("Index out of range (pars): case 2 indexes: "+std::to_string(indexes[idx_iter])+" versus "+std::to_string(parameters.size()));
 #endif
